Start
“seaLevels 10” should show a graph of the sea levels in 10 years of
these six major US cities. One should be able to visualize the change
in seaLevels and see that they are increasing.

To further this code, I would like to start with a graphic of where the
levels are at currently and then have you input and see the change.

For some reason seaLevels is not being defined and I do not know why.
Please let me know if you are around on Monday(!!) I will message you
tomorrow. Sorry that I always do these Sunday nights, I work weekends.
I really care and want to be less challenged!